Ensure we set job status to failing after retries finish
The `retry_count` number that Sidekiq makes available to retried jobs
can be a little confusing. It's actually the number of retries that we
have completed, minus one. It's easiest to understand via an example.

Given a job with `retries: 3`, we'll see the following pattern:

- initial run (fails), middleware was called with `msg['retry_count']`
  not present
- first retry (fails), middleware was called with `msg['retry_count']
  == 0`
- second retry (fails), middleware was called with `msg['retry_count']
  == 1`
- third retry (fails), middleware was called with `msg['retry_count']
  == 2`

The third retry is the final one, after which we _should_ be setting
the status to `:failed`, but the retry count at that point is `2`,
which is less than the value `retry_attempts_from` will return (`3`),
and so the status remained `:retrying`.

Part of the reason this is confusing is because this `retry_count`
value feels like it should reflect the current state of the job after
it fails, but it's actually the same value right even before the job
executes. The right way of thinking about it is something like "this
was retry attempt #0", and the first retry is indexed `0`, somewhat
confusingly.

This behaviour appears to be present in Sidekiq 4 and 5, so to make
things work again, we can add an offset when running under those
versions, to make the comparison work and correctly set the status to
`:failed` after the final retry does not succeed.

I've tried to extract this into private methods that make things a bit
clearer.