Don't time out before jobs get retried
After introducing a job that does a retry, we'd occasionally see the
tests fail because of a timeout. This is because of the random factor
involved in the delay of the retry. The retry amount is calculate as

    (count ** 4) + 15 + (rand(30)*(count+1))

and so the maximum delay for the first try would be

    (0 ** 4) + 15 + (29*(0+1)) == 44

This is greater than the timeout we use in tests, causing the build to
fail. Without a way of configuring the exponential backoff, the only
way to fix this while still including tests that do retry a job is by
increasing the timeout.

I've picked 60 as a fairly arbitrary round number that's more than 44.