resolve job classes that have not been loaded yet
If you pass a string for the job class, and the constant has already been loaded, the world is a happy place.  However, if the constant has not yet been loaded (lazy loading in Rails for example) then this breaks.  This fix attempts to autoload the constant to fix this issue.